// taskflow-frontend/src/app/core/models/auth.model.ts
import { Role } from './enums';

export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  accessToken: string;
  refreshToken: string;
  expiresIn: number;
  role: Role;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
}

export interface RegisterResponse {
  id: number;
  name: string;
  email: string;
  role: Role;
}

export interface AuthUser {
  userId: number;
  email: string;
  role: Role;
}